Move t related util functions to Keep lib
