+++ solidity-v1/dashboard/src/constants/constants.js
@@ -329,6 +329,3 @@ export const UNDELEGATE_STAKE_CALENDAR_EVENT = {
   // eslint-disable-next-line no-template-curly-in-string
   location: "https://dashboard.keep.network/${address}/overview",
 }
-
-// You can check the ratio here: https://etherscan.io/address/0xE47c80e8c23f6B4A1aE41c34837a0599D5D16bb0#readContract
-export const KEEP_TO_T_EXCHANGE_RATE_IN_WEI = 4783188631255016
